Add models and register models to admin
In terms of extention, Photo model should not be applied to Board app
It needs to be added into Photo app so other app can use it.

Here are the models that would be contained for now
: Category, Board, Comment, Tag

In addition, the User app should contains customised User model and its manager.