Remove redundancy
As the default value of permission_classes is permissions.AllowAny,
removed the line that specify permission classes on default swagger urls.

If `permissions.AllowAny` is set, it will cause a mypy error
: error: Argument "permission_classes" to "get_schema_view" has incompatible type
  "Tuple[Type[AllowAny]]"; expected "Optional[Tuple[str]]"

This, `Optional[Tuple[str]], means: str 'rest_framework.permissions.AllowAny'
in settings.py of `rest_framework` that is redundancy.